Add method to RelayFrame to reveal Arg2 offset
Add a method "Arg2EndOffset" to RelayFrame interface so that we could
know how much space Arg2 is used. This is the first step to know if it's
possible to add more methods in Relay code to allow Relayer to inject
more headers or read the headers to decide which peer to forward to.

NOTE: this is a breaking change to RelayFrame interface.

Related to #745. Will use this change to know Arg2 usage.