introspection: Introspect any channel by ID
Currently, the introspection endpoint only lets you (fully) introspect
the channel that received the request. This restricts introspection to
only work for channels that listen, but there are some services that
create multiple channels: 1 for listening, and then 1 per client.

Since the per-client channels may not be listening, we can only get
limited information on their state. Since we already have unique IDs
for each channel (and this is exposed via the `includeOtherChannels`),
provide a way to introspect a specific channel by ID.

This is backwards compatible -- if no ID is specified, the current
channel is introspected as it is today.