relay: Log send buffer usage and size when dropping calls
Ref T2984301

We drop calls when the sendCh is full, which is typically because
writeFrames is not reading from sendCh quickly enough. This could
be because of unfortunate scheduling, or because the send buffer is
full and write syscalls were blocked. Logging the send buffer usage
will help determine which is the cause of dropped calls.

We can get the socket send buffer usage with ioctl, and buffer size
using getsockopt.

ioctl with SIOCOUTQ returns the number of bytes in the send buffer:
See https://linux.die.net/man/7/tcp for details
> Returns the amount of unsent data in the socket send queue.