WIP: Add timeout option for writing frames to sendCh
Ref T2984301

Instead of dropping calls immediately if the sendCh is full, add
a timeout option which will wait for a specified duration of time
for the sendCh to have space.

This helps in the case that the client is actively reading from the
socket, or in the case that the writeFrames goroutine is actively
writing to the socket, and applies some backpressure to the backend
generating large responses.

This does add latency to process all responses from that backend, so
if there's other clients who have responses in the backend socket's
buffer, they will wait for the timeout as well.