relay: Log send buffer usage and size when dropping calls (#768)
Ref T5413559

We drop calls when the sendCh is full, which is typically because
writeFrames is not reading from sendCh quickly enough. This could
be because of unfortunate scheduling, or because the send buffer is
full and write syscalls were blocked. Logging the send buffer usage
will help determine which is the cause of dropped calls.

We can get the socket send buffer usage with ioctl, and buffer size
using getsockopt.

ioctl with SIOCOUTQ returns the number of bytes in the send buffer:
See https://linux.die.net/man/7/tcp for details
> Returns the amount of unsent data in the socket send queue.

Extend the introspection output to contain the same buffer sizes as
well as the sendCh length.

Example log message fields added to the log:
```
   "destConnSendBufferCurrent":3994463,
   "destConnSendBufferLimit":4194304
```