relay: Limit connection timeouts to avoid stalls
Ref T5890079

If we need to connect in the relay path, we're basically stalling the
connection (all calls on the connection will be backed up behind the
connection). If a call has a large timeout (e.g., 5 seconds), and the
backend is unavailable, leading to a timeout, then all calls behind will
likely be timed out by the time we've started processing them.

Limit the connection timeout to a shorter value than the call timeout
since connection time is known at the infrastructural layer.