Fix flaky TestRelayBasedSweep test
TestRelayBasedSweep verifies that the relay closes idle connections.
On CI, we saw failures where idle connections were not closed. It took
a while to repro (needed a strategically placed `time.Sleep` at the star
of `pollerLoop`), but it turned out to be a incorrect assumption in the
test.

The test assumes that `relayTicker` is only used by the relay to create
a ticker for the idle sweeper. However, `TestServer` actually
creates a server channel + relay channel with the same opts. In our
case, we don't use the server, but it's already been created.

This causes issues as when we do `relayTicker.Tick()`, it's a race
on which channel's idle sweeper will run, and if the server channel
wins the race, then the relay's idle sweeper doesn't run and doesn't
close connections as expected.

To fix this, add a new option to the TestServer to disable the server,
and do some refactorings to ensure that `ts.Server()` and `ts.Relay()`
work and are only used when they are actually enabled.

Fixes #689.