relay: Explicitly track originator in relayItem
Ref T6369485

We intended to store whether a relayItem represents the
originator (caller), but the `local` bool was unused.

Rename for consistency to `isOriginator`, set it correctly, and use
this value instead of `item.call != nil`.

This prefactor removes an assumption that `item.call` is only set
on the originator, since we want to store `item.call` on the dest
item for tracking transferred bytes for T6369485.

This change has no behaviour changes, it is a pure refactor.