relay: Fix races between timeout & call completion
Fixes #808.

As described in #808, we have a race between deletions releasing a timer
and a concurrent goroutine holding on to a reference to `item` which has
a now-stale reference to the timeout that is attempted to be stopped.

To avoid this race, instead of stopping timeouts with no lock held, stop
timeouts while holding an RLock in Get. This ensures that once we delete
an item and release the timeout, no other goroutines will call
timeout.Stop since that cannot happen concurrently (due to locks), and
the concurrent goroutine will either run before (in which case they will
stop the timeout) or after (in which case the element will not be
found).

This same logic is used anywhere we stop timeouts (avoiding similar
races in other codepaths).

With this fix, TestRelayRaceCompletionAndTimeout does not trigger any
races, even with >1000 runs.