relay: Send error frame to caller on fragmented arg2
Currently, if a call fails due to a fragmented arg2 (e.g., arg2 does not
fit in the first frame as it's too large), we drop the frame and log
an error internally, but no error is returned to the caller, so they
see a timeout.

Instead of a timeout, fail the relay item, which triggers an error frame
to be sent to the caller, so they receive mroe information on why the
call failed (and they receive an immediate error instead of waiting for
the timeout to expire).

This also improves other error frames sent back with the same
reason that is used in metrics, although we still have a generic
"frame was not sent to remote side" which could be replaced with
much more specific errors.