Drop logs when idle connections are closed (#820)
We currently have 2 logs whenever an idle connection is closed:
 * "Closing idle inbound connection"
 * "Connection closing." with reason="Idle connection closed"

The first log doesn't provide much value since the closing log already gives us
a detailed reason. Since it's redundant, drop the first log, and pass any fields
that only the first log had to the `close`, so they are included in the second log.