relay: Fix overflow when reading large service (#827)
Service names are expected to end at (_serviceNameIndex + length).
However, the length of the service is read as a byte, and
_serviceNameIndex is a constant (untyped). This results in the above
addition happening as a byte instead of an int, so it ends up
overflowing when _serviceNameIndex + length > 255. This overflow
can result in the end index being incorrect (pointing to before the
start index), which results in a panic:
```
panic: runtime error: slice bounds out of range [31:20]

goroutine 142 [running]:
github.com/uber/tchannel-go.(*lazyCallReq).Service(...)
        [..]/tchannel-go/relay_messages.go:187
[...]
```

To fix this, we need to do all operations using ints. To enforce this,
update all index constants to be typed as ints, as arithmetic operations
between 2 different types is disallowed, and fails on byte + int.

Update the test frame generator to support arbitrary service names
and adapt the frame length based on the contents vs a hardcoded value.