Allow method registration if handler implements Register (#832)
Currently, channel/subchannel only allow Register if the type
of the handler is the inbuilt channelHandler or handlerMap types.

This is quite inflexible, e.g., as we add new handler types in #831,
we still want to support registration without extending the type checks.

Instead of looking for a specific type, we can make the logic cleaner
by looking for the Register method. This also makes the channel's
Register method, and the subchannel's Register method more consistent.

This change allows users to specify their own custom handlers that can
support registration (via the same Register method).