refactor callRes test frame construction (#834)
Prefactor for #830

I'd like to refactor the code for constructing callRes frames so that we can easily add more scenarios such as malformed args or different arg schemes. This is important since arg2 iteration is supported only for thrift calls, and we will be doing some error checking in newLazyCallRes which can be complicated to test with the current code.