Fix memory leak in the relayer resulting from dropped frames
Recently it was discovered that a number of code paths in
the relayer would result in frame leaks, where after a relay
decision (either send to `Connection.writeFrames()` for sending
or dropped) the frame isn't returned to the frame pool,
resulting in a memory leak.

With the previous PRs #847 and #852, we are now in a position
to address the leaks in each code path. This change fixes the
leaks in `handleNonCallReq()` as detected by the existing tests.