Add option to check frame leaks after each test with WtihTestServer
With the realization that some code paths aren't properly releasing frames
(e.g. #845), I'd like to add an option to WithTestServer to optionally check
for frame leaks along with each test.

As RecordingFramePool isn't an actual frame pool, we can't simply replace
the frame pool in all of our tests - we'd miss signals on how the code
is actually run in production. As such, we'll need to run the same test again
with RecordingFramePool, which will more than double the time it takes
to run a test.

To avoid an outsized impact to test duration, make the option opt-in so that
we can selectively onboard tests to this check.