Fix flaky TestRelayConnection (#851)
See #848 for context. This fixes the same issue but instead of waiting
for the inbound connection, we use Ping, which ensures the connection is
added to the channels on both the inbound/outbound side, and then
returns the pingRes.

This is similar to other fixes in the past, e.g., #440, #433.