Fix flaky TestIdleSweepIgnoresConnectionsWithCalls
Due to a race between the idle connection sweeper and message
exchange, the test may end up getting a false positive if the
exchange shutdown happened after the last sweep has been run.

To mitigate this, add an extra wait for the number of exchanges
for all peers to drop to zero.