Fix flaky TestLargeTimeoutsAreClamped
TestLargeTimeoutsAreClamped tests that timeouts for relayed calls are "clamped"
by:
- Setting up a backend with a blocking handler
- Setting up a relay with a max timeout of `clampTTL`
- Making a call from a client with a timeout of `longTTL`
- Checking that the call returns with a timeout error within 10x `clampTTL`

This worked before, but since the migration to github actions, this test has
become very flaky, likely due to heavier load than TravisCI. The flakiness arises
from the fact that since it's a "blackbox" style test, system noise can easily
factor into the timing, causing e.g. the latency in the relay to increase significantly
under heavy load and breaching the time limit of 10x `clampTTL`.

This change bumps the timeout in the final check to `longTTL/2` to account for
latencies on the order of 100ms. The purpose here becomes checking that we timed out
before `longTTL` is reached, rather than verifying that we're clamping exactly
at `clampTTL`.