Fix frame leaks in TestRelayStalledConnection
Due to a race between the message exchange and relayer,
a subtle frame leak would only show up when the message
exchange was shut down (receiver closed the connection)
before the relayer stopped sending frames.

This leak is due to 2 reasons:
- When an exchange is shut down before the relayer has finished relaying,
  the frame can no longer be forwarded. But since we didn't treat that
  as an error, we also didn't release the frame in question.
- When the message exchange shuts down, it doesn't drain the receive
  buffer.

This change fixes the issues above. The test is tested with
100s of repeated runs to ensure it is no longer leaking frames
intermittently.