Optionally send cancelled frames when context is canceled (#890)
Replaces #817

Add a pair of new options, `SendCancelOnContextCanceled` (sender) and `PropagateCancel`
(receiver) to support sending cancel frames when a context is detected to be canceled on the
request sender side.

Cancellations often happen while waiting to read a response, so we can rely on the
reader goroutine waiting on the message exchange to notice the cancellation and
send the cancel message to the remote side vs adding a separate goroutine.

Older servers and relays without this change will log Received unexpected frame
but otherwise be unaffected by the unknown frame. However, there's no point enabling
cancellations if the remote side doesn't support them first.