Add component tag to tracing spans (#923)
This PR is making a minor change to add span component tag to tracing spans. According to the [OpneTracing sematic conventions](https://github.com/opentracing/specification/blob/master/semantic_conventions.md), the component tag indicates the library that created the spans. At Uber, It is quite useful for the tracing team to understand the span creation distribution among libraries in order to optimize the operational cost.
> The software package, framework, library, or module that generated the associated Span. E.g., "grpc", "django", "JDBI".